Moved ContextProvider module into Web3 module; added methods to wallet context providers
