+++ src/Web3/Web3Connect/ContextProvider/WalletContextProvider.jsx
@@ -6,8 +6,15 @@ function WalletContextProvider(props) {
   const { children } = props;
   const [walletProvider, setWalletProvider] = useState();
 
+  const getAddresses = () => {
+    return walletProvider._addresses;
+  }
+
+  const getChainId = () => {
+    return walletProvider.chainId;
+  }
   return (
-    <WalletContext.Provider value={{ walletProvider, setWalletProvider }}>
+    <WalletContext.Provider value={{ walletProvider, setWalletProvider, getAddresses, getChainId }}>
       {children}
     </WalletContext.Provider>
   );
