+++ src/Web3/Web3Connect/index.js
@@ -1,2 +1,3 @@
 export * from "./Component";
 export * from "./Utils";
+export * from "./ContextProvider";
